fix(engine-server): reflect global HTML attributes
