+++ packages/@lwc/engine-server/src/__tests__/fixtures/global-html-attributes/expected.html
@@ -3,74 +3,74 @@
     <x-child accesskey="tata" aria-label="titi" dir="auto" draggable="false" id="tutu" lang="jp" role="scrollbar" spellcheck="false" tabindex="0" title="tete">
       <template shadowrootmode="open">
         <div>
-          accessKey: 
+          accessKey: tata
         </div>
         <div>
           ariaLabel: titi
         </div>
         <div>
-          dir: 
+          dir: auto
         </div>
         <div>
-          draggable: 
+          draggable: false
         </div>
         <div>
-          hidden: 
+          hidden: false
         </div>
         <div>
-          id: 
+          id: tutu
         </div>
         <div>
-          lang: 
+          lang: jp
         </div>
         <div>
           role: scrollbar
         </div>
         <div>
-          spellcheck: 
+          spellcheck: true
         </div>
         <div>
-          tabIndex: 
+          tabIndex: 0
         </div>
         <div>
-          title: 
+          title: tete
         </div>
       </template>
     </x-child>
     <x-child accesskey="tata" aria-label="titi" dir="auto" draggable="true" hidden id="tutu" lang="jp" role="scrollbar" spellcheck="false" tabindex="0" title="tete">
       <template shadowrootmode="open">
         <div>
-          accessKey: 
+          accessKey: tata
         </div>
         <div>
           ariaLabel: titi
         </div>
         <div>
-          dir: 
+          dir: auto
         </div>
         <div>
-          draggable: 
+          draggable: true
         </div>
         <div>
-          hidden: 
+          hidden: true
         </div>
         <div>
-          id: 
+          id: tutu
         </div>
         <div>
-          lang: 
+          lang: jp
         </div>
         <div>
           role: scrollbar
         </div>
         <div>
-          spellcheck: 
+          spellcheck: true
         </div>
         <div>
-          tabIndex: 
+          tabIndex: 0
         </div>
         <div>
-          title: 
+          title: tete
         </div>
       </template>
     </x-child>
