+++ packages/@lwc/shared/src/index.ts
@@ -17,6 +17,7 @@ export * from './html-escape';
 export * from './meta';
 export * from './namespaces';
 export * from './overridable-hooks';
+export * from './attr-reflection';
 export * from './static-part-tokens';
 export * from './style';
 export * from './signals';
